feat: relocate modules
