+++ bcs-app/backend/helm/app/views.py
@@ -37,6 +37,7 @@ from rest_framework.serializers import Serializer
 from backend.accounts import bcs_perm
 from backend.components import paas_cc
 from backend.components.bcs import k8s
+from backend.container_service.misc.bke_client.client import BCSClusterCredentialsNotFound, BCSClusterNotFound
 from backend.helm.app.repo import get_or_create_private_repo
 from backend.helm.app.serializers import FilterNamespacesSLZ
 from backend.helm.app.utils import get_helm_dashboard_path
@@ -45,7 +46,6 @@ from backend.helm.helm.models.chart import ChartVersion
 from backend.helm.helm.providers.repo_provider import add_plain_repo, add_platform_public_repos, add_repo
 from backend.helm.helm.serializers import ChartVersionSLZ
 from backend.helm.permissions import check_cluster_perm
-from backend.helm.toolkits.bke_client.client import BCSClusterCredentialsNotFound, BCSClusterNotFound
 from backend.helm.toolkits.diff import parser
 from backend.kube_core.toolkit.dashboard_cli.exceptions import DashboardError, DashboardExecutionError
 from backend.resources.namespace.constants import K8S_SYS_PLAT_NAMESPACES
