+++ bcs-app/backend/utils/client.py
@@ -20,7 +20,7 @@ from kubernetes.client.rest import ApiException
 from rest_framework.exceptions import APIException
 
 from backend.components import bcs
-from backend.helm.toolkits.bke_client import BCSClusterClient
+from backend.container_service.misc.bke_client import BCSClusterClient
 from backend.kube_core.toolkit.dashboard_cli import DashboardClient
 from backend.kube_core.toolkit.kubectl import KubectlClusterClient
 from backend.utils.error_codes import error_codes
